Split date picker 2 code to multiple files
