+++ lib/calendar_date_picker2.dart
@@ -1,5 +1,5 @@
 library calendar_date_picker2;
 
-export 'src/models/index.dart';
-export 'src/utils/index.dart';
-export 'src/widgets/index.dart';
+export 'src/models/models.dart';
+export 'src/utils/utils.dart';
+export 'src/widgets/widgets.dart';
