+++ lib/src/models/models.dart
@@ -1,0 +1,1 @@
+export 'calendar_date_picker2_config.dart';
