+++ lib/src/widgets/header.dart
@@ -1,0 +1,151 @@
+import 'package:calendar_date_picker2/src/constants.dart';
+import 'package:flutter/material.dart';
+
+import '../../calendar_date_picker2.dart';
+
+/// A button that used to toggle the [DatePickerMode] for a date picker.
+///
+/// This appears above the calendar grid and allows the user to toggle the
+/// [DatePickerMode] to display either the calendar view or the year list.
+class DatePickerModeToggleButton extends StatefulWidget {
+  const DatePickerModeToggleButton({
+    Key? key,
+    required this.mode,
+    required this.title,
+    required this.onTitlePressed,
+    required this.config,
+  }) : super(key: key);
+
+  /// The current display of the calendar picker.
+  final DatePickerMode mode;
+
+  /// The text that displays the current month/year being viewed.
+  final String title;
+
+  /// The callback when the title is pressed.
+  final VoidCallback onTitlePressed;
+
+  /// The calendar configurations
+  final CalendarDatePicker2Config config;
+
+  @override
+  DatePickerModeToggleButtonState createState() =>
+      DatePickerModeToggleButtonState();
+}
+
+class DatePickerModeToggleButtonState extends State<DatePickerModeToggleButton>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      value: widget.mode == DatePickerMode.year ? 0.5 : 0,
+      upperBound: 0.5,
+      duration: const Duration(milliseconds: 200),
+      vsync: this,
+    );
+  }
+
+  @override
+  void didUpdateWidget(DatePickerModeToggleButton oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (oldWidget.mode == widget.mode) {
+      return;
+    }
+
+    if (widget.mode == DatePickerMode.year) {
+      _controller.forward();
+    } else {
+      _controller.reverse();
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
+    final TextTheme textTheme = Theme.of(context).textTheme;
+    final Color controlColor = colorScheme.onSurface.withOpacity(0.60);
+    var datePickerOffsetPadding = C2Constants.monthNavButtonsWidth;
+    if (widget.config.centerAlignModePicker == true) {
+      datePickerOffsetPadding /= 2;
+    }
+
+    return Container(
+      padding: widget.config.centerAlignModePicker == true
+          ? EdgeInsets.zero
+          : const EdgeInsetsDirectional.only(start: 16, end: 4),
+      height: (widget.config.controlsHeight ?? C2Constants.subHeaderHeight),
+      child: Row(
+        children: <Widget>[
+          if (widget.mode == DatePickerMode.day &&
+              widget.config.centerAlignModePicker == true)
+            // Give space for the prev/next month buttons that are underneath this row
+            SizedBox(width: datePickerOffsetPadding),
+          Flexible(
+            child: Semantics(
+              label: MaterialLocalizations.of(context).selectYearSemanticsLabel,
+              excludeSemantics: true,
+              button: true,
+              child: SizedBox(
+                height: (widget.config.controlsHeight ??
+                    C2Constants.subHeaderHeight),
+                child: InkWell(
+                  onTap: widget.config.disableModePicker == true
+                      ? null
+                      : widget.onTitlePressed,
+                  child: Padding(
+                    padding: EdgeInsets.symmetric(
+                        horizontal: widget.config.centerAlignModePicker == true
+                            ? 0
+                            : 8),
+                    child: Row(
+                      mainAxisAlignment:
+                          widget.config.centerAlignModePicker == true
+                              ? MainAxisAlignment.center
+                              : MainAxisAlignment.start,
+                      children: <Widget>[
+                        Flexible(
+                          child: Text(
+                            widget.title,
+                            overflow: TextOverflow.ellipsis,
+                            style: widget.config.controlsTextStyle ??
+                                textTheme.titleSmall?.copyWith(
+                                  color: controlColor,
+                                ),
+                          ),
+                        ),
+                        widget.config.disableModePicker == true
+                            ? const SizedBox()
+                            : RotationTransition(
+                                turns: _controller,
+                                child: widget.config.customModePickerIcon ??
+                                    Icon(
+                                      Icons.arrow_drop_down,
+                                      color: widget.config.controlsTextStyle
+                                              ?.color ??
+                                          controlColor,
+                                    ),
+                              ),
+                      ],
+                    ),
+                  ),
+                ),
+              ),
+            ),
+          ),
+          if (widget.mode == DatePickerMode.day)
+            // Give space for the prev/next month buttons that are underneath this row
+            SizedBox(width: datePickerOffsetPadding),
+        ],
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+}
