+++ lib/src/widgets/month_picker.dart
@@ -1,0 +1,380 @@
+import 'package:calendar_date_picker2/src/constants.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
+
+import '../../calendar_date_picker2.dart';
+import 'day_picker.dart';
+import 'focus_date.dart';
+
+T? _ambiguate<T>(T? value) => value;
+
+class C2MonthPicker extends StatefulWidget {
+  /// Creates a month picker.
+  const C2MonthPicker({
+    required this.config,
+    required this.initialMonth,
+    required this.selectedDates,
+    required this.onChanged,
+    required this.onDisplayedMonthChanged,
+    Key? key,
+  }) : super(key: key);
+
+  /// The calendar configurations
+  final CalendarDatePicker2Config config;
+
+  /// The initial month to display.
+  final DateTime initialMonth;
+
+  /// The currently selected dates.
+  ///
+  /// Selected dates are highlighted in the picker.
+  final List<DateTime?> selectedDates;
+
+  /// Called when the user picks a day.
+  final ValueChanged<DateTime> onChanged;
+
+  /// Called when the user navigates to a new month.
+  final ValueChanged<DateTime> onDisplayedMonthChanged;
+
+  @override
+  C2MonthPickerState createState() => C2MonthPickerState();
+}
+
+class C2MonthPickerState extends State<C2MonthPicker> {
+  final GlobalKey _pageViewKey = GlobalKey();
+  late DateTime _currentMonth;
+  late PageController _pageController;
+  late MaterialLocalizations _localizations;
+  late TextDirection _textDirection;
+  Map<ShortcutActivator, Intent>? _shortcutMap;
+  Map<Type, Action<Intent>>? _actionMap;
+  late FocusNode _dayGridFocus;
+  DateTime? _focusedDay;
+
+  @override
+  void initState() {
+    super.initState();
+    _currentMonth = widget.initialMonth;
+    _pageController = PageController(
+        initialPage:
+            DateUtils.monthDelta(widget.config.firstDate, _currentMonth));
+    _shortcutMap = const <ShortcutActivator, Intent>{
+      SingleActivator(LogicalKeyboardKey.arrowLeft):
+          DirectionalFocusIntent(TraversalDirection.left),
+      SingleActivator(LogicalKeyboardKey.arrowRight):
+          DirectionalFocusIntent(TraversalDirection.right),
+      SingleActivator(LogicalKeyboardKey.arrowDown):
+          DirectionalFocusIntent(TraversalDirection.down),
+      SingleActivator(LogicalKeyboardKey.arrowUp):
+          DirectionalFocusIntent(TraversalDirection.up),
+    };
+    _actionMap = <Type, Action<Intent>>{
+      NextFocusIntent:
+          CallbackAction<NextFocusIntent>(onInvoke: _handleGridNextFocus),
+      PreviousFocusIntent: CallbackAction<PreviousFocusIntent>(
+          onInvoke: _handleGridPreviousFocus),
+      DirectionalFocusIntent: CallbackAction<DirectionalFocusIntent>(
+          onInvoke: _handleDirectionFocus),
+    };
+    _dayGridFocus = FocusNode(debugLabel: 'Day Grid');
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _localizations = MaterialLocalizations.of(context);
+    _textDirection = Directionality.of(context);
+  }
+
+  @override
+  void didUpdateWidget(C2MonthPicker oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.initialMonth != oldWidget.initialMonth &&
+        widget.initialMonth != _currentMonth) {
+      // We can't interrupt this widget build with a scroll, so do it next frame
+      // Add workaround to fix Flutter 3.0.0 compiler issue
+      // https://github.com/flutter/flutter/issues/103561#issuecomment-1125512962
+      // https://github.com/flutter/website/blob/3e6d87f13ad2a8dd9cf16081868cc3b3794abb90/src/development/tools/sdk/release-notes/release-notes-3.0.0.md#your-code
+      _ambiguate(WidgetsBinding.instance)!.addPostFrameCallback(
+        (Duration timeStamp) => _showMonth(widget.initialMonth, jump: true),
+      );
+    }
+  }
+
+  @override
+  void dispose() {
+    _pageController.dispose();
+    _dayGridFocus.dispose();
+    super.dispose();
+  }
+
+  void _handleDateSelected(DateTime selectedDate) {
+    _focusedDay = selectedDate;
+    widget.onChanged(selectedDate);
+  }
+
+  void _handleMonthPageChanged(int monthPage) {
+    setState(() {
+      final DateTime monthDate =
+          DateUtils.addMonthsToMonthDate(widget.config.firstDate, monthPage);
+      if (!DateUtils.isSameMonth(_currentMonth, monthDate)) {
+        _currentMonth = DateTime(monthDate.year, monthDate.month);
+        widget.onDisplayedMonthChanged(_currentMonth);
+        if (_focusedDay != null &&
+            !DateUtils.isSameMonth(_focusedDay, _currentMonth)) {
+          // We have navigated to a new month with the grid focused, but the
+          // focused day is not in this month. Choose a new one trying to keep
+          // the same day of the month.
+          _focusedDay = _focusableDayForMonth(_currentMonth, _focusedDay!.day);
+        }
+        SemanticsService.announce(
+          _localizations.formatMonthYear(_currentMonth),
+          _textDirection,
+        );
+      }
+    });
+  }
+
+  /// Returns a focusable date for the given month.
+  ///
+  /// If the preferredDay is available in the month it will be returned,
+  /// otherwise the first selectable day in the month will be returned. If
+  /// no dates are selectable in the month, then it will return null.
+  DateTime? _focusableDayForMonth(DateTime month, int preferredDay) {
+    final int daysInMonth = DateUtils.getDaysInMonth(month.year, month.month);
+
+    // Can we use the preferred day in this month?
+    if (preferredDay <= daysInMonth) {
+      final DateTime newFocus = DateTime(month.year, month.month, preferredDay);
+      if (_isSelectable(newFocus)) return newFocus;
+    }
+
+    // Start at the 1st and take the first selectable date.
+    for (int day = 1; day <= daysInMonth; day++) {
+      final DateTime newFocus = DateTime(month.year, month.month, day);
+      if (_isSelectable(newFocus)) return newFocus;
+    }
+    return null;
+  }
+
+  /// Navigate to the next month.
+  void _handleNextMonth() {
+    if (!_isDisplayingLastMonth) {
+      _pageController.nextPage(
+        duration: C2Constants.monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  /// Navigate to the previous month.
+  void _handlePreviousMonth() {
+    if (!_isDisplayingFirstMonth) {
+      _pageController.previousPage(
+        duration: C2Constants.monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  /// Navigate to the given month.
+  void _showMonth(DateTime month, {bool jump = false}) {
+    final int monthPage = DateUtils.monthDelta(widget.config.firstDate, month);
+    if (jump) {
+      _pageController.jumpToPage(monthPage);
+    } else {
+      _pageController.animateToPage(
+        monthPage,
+        duration: C2Constants.monthScrollDuration,
+        curve: Curves.ease,
+      );
+    }
+  }
+
+  /// True if the earliest allowable month is displayed.
+  bool get _isDisplayingFirstMonth {
+    return !_currentMonth.isAfter(
+      DateTime(widget.config.firstDate.year, widget.config.firstDate.month),
+    );
+  }
+
+  /// True if the latest allowable month is displayed.
+  bool get _isDisplayingLastMonth {
+    return !_currentMonth.isBefore(
+      DateTime(widget.config.lastDate.year, widget.config.lastDate.month),
+    );
+  }
+
+  /// Handler for when the overall day grid obtains or loses focus.
+  void _handleGridFocusChange(bool focused) {
+    setState(() {
+      if (focused && _focusedDay == null && widget.selectedDates.isNotEmpty) {
+        if (DateUtils.isSameMonth(widget.selectedDates[0], _currentMonth)) {
+          _focusedDay = widget.selectedDates[0];
+        } else if (DateUtils.isSameMonth(
+            widget.config.currentDate, _currentMonth)) {
+          _focusedDay = _focusableDayForMonth(
+              _currentMonth, widget.config.currentDate.day);
+        } else {
+          _focusedDay = _focusableDayForMonth(_currentMonth, 1);
+        }
+      }
+    });
+  }
+
+  /// Move focus to the next element after the day grid.
+  void _handleGridNextFocus(NextFocusIntent intent) {
+    _dayGridFocus.requestFocus();
+    _dayGridFocus.nextFocus();
+  }
+
+  /// Move focus to the previous element before the day grid.
+  void _handleGridPreviousFocus(PreviousFocusIntent intent) {
+    _dayGridFocus.requestFocus();
+    _dayGridFocus.previousFocus();
+  }
+
+  /// Move the internal focus date in the direction of the given intent.
+  ///
+  /// This will attempt to move the focused day to the next selectable day in
+  /// the given direction. If the new date is not in the current month, then
+  /// the page view will be scrolled to show the new date's month.
+  ///
+  /// For horizontal directions, it will move forward or backward a day (depending
+  /// on the current [TextDirection]). For vertical directions it will move up and
+  /// down a week at a time.
+  void _handleDirectionFocus(DirectionalFocusIntent intent) {
+    setState(() {
+      if (_focusedDay != null) {
+        final nextDate = _nextDateInDirection(_focusedDay!, intent.direction);
+        if (nextDate != null) {
+          _focusedDay = nextDate;
+          if (!DateUtils.isSameMonth(_focusedDay, _currentMonth)) {
+            _showMonth(_focusedDay!);
+          }
+        }
+      } else {
+        _focusedDay ??= widget.initialMonth;
+      }
+    });
+  }
+
+  static const Map<TraversalDirection, int> _directionOffset =
+      <TraversalDirection, int>{
+    TraversalDirection.up: -DateTime.daysPerWeek,
+    TraversalDirection.right: 1,
+    TraversalDirection.down: DateTime.daysPerWeek,
+    TraversalDirection.left: -1,
+  };
+
+  int _dayDirectionOffset(
+      TraversalDirection traversalDirection, TextDirection textDirection) {
+    // Swap left and right if the text direction if RTL
+    if (textDirection == TextDirection.rtl) {
+      if (traversalDirection == TraversalDirection.left) {
+        traversalDirection = TraversalDirection.right;
+      } else if (traversalDirection == TraversalDirection.right) {
+        traversalDirection = TraversalDirection.left;
+      }
+    }
+    return _directionOffset[traversalDirection]!;
+  }
+
+  DateTime? _nextDateInDirection(DateTime date, TraversalDirection direction) {
+    final TextDirection textDirection = Directionality.of(context);
+    DateTime nextDate = DateUtils.addDaysToDate(
+        date, _dayDirectionOffset(direction, textDirection));
+    while (!nextDate.isBefore(widget.config.firstDate) &&
+        !nextDate.isAfter(widget.config.lastDate)) {
+      if (_isSelectable(nextDate)) {
+        return nextDate;
+      }
+      nextDate = DateUtils.addDaysToDate(
+          nextDate, _dayDirectionOffset(direction, textDirection));
+    }
+    return null;
+  }
+
+  bool _isSelectable(DateTime date) {
+    return widget.config.selectableDayPredicate?.call(date) ?? true;
+  }
+
+  Widget _buildItems(BuildContext context, int index) {
+    final DateTime month =
+        DateUtils.addMonthsToMonthDate(widget.config.firstDate, index);
+    return C2DayPicker(
+      key: ValueKey<DateTime>(month),
+      selectedDates: (widget.selectedDates..removeWhere((d) => d == null))
+          .cast<DateTime>(),
+      onChanged: _handleDateSelected,
+      config: widget.config,
+      displayedMonth: month,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final Color controlColor =
+        Theme.of(context).colorScheme.onSurface.withOpacity(0.60);
+
+    return Semantics(
+      child: Column(
+        children: <Widget>[
+          Container(
+            padding: widget.config.centerAlignModePicker != true
+                ? const EdgeInsetsDirectional.only(start: 16, end: 4)
+                : const EdgeInsetsDirectional.only(start: 8, end: 8),
+            height:
+                (widget.config.controlsHeight ?? C2Constants.subHeaderHeight),
+            child: Row(
+              children: <Widget>[
+                if (widget.config.centerAlignModePicker != true) const Spacer(),
+                IconButton(
+                  icon: widget.config.lastMonthIcon ??
+                      const Icon(Icons.chevron_left),
+                  color: controlColor,
+                  tooltip: _isDisplayingFirstMonth
+                      ? null
+                      : _localizations.previousMonthTooltip,
+                  onPressed:
+                      _isDisplayingFirstMonth ? null : _handlePreviousMonth,
+                ),
+                if (widget.config.centerAlignModePicker == true) const Spacer(),
+                IconButton(
+                  icon: widget.config.nextMonthIcon ??
+                      const Icon(Icons.chevron_right),
+                  color: controlColor,
+                  tooltip: _isDisplayingLastMonth
+                      ? null
+                      : _localizations.nextMonthTooltip,
+                  onPressed: _isDisplayingLastMonth ? null : _handleNextMonth,
+                ),
+              ],
+            ),
+          ),
+          Expanded(
+            child: FocusableActionDetector(
+              shortcuts: _shortcutMap,
+              actions: _actionMap,
+              focusNode: _dayGridFocus,
+              onFocusChange: _handleGridFocusChange,
+              child: C2FocusedDate(
+                date: _dayGridFocus.hasFocus ? _focusedDay : null,
+                child: PageView.builder(
+                  key: _pageViewKey,
+                  controller: _pageController,
+                  itemBuilder: _buildItems,
+                  itemCount: DateUtils.monthDelta(
+                          widget.config.firstDate, widget.config.lastDate) +
+                      1,
+                  onPageChanged: _handleMonthPageChanged,
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
