+++ lib/src/widgets/year_picker.dart
@@ -1,0 +1,239 @@
+import 'package:calendar_date_picker2/calendar_date_picker2.dart';
+import 'package:calendar_date_picker2/src/constants.dart';
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+import 'dart:math' as math;
+
+import 'package:flutter/rendering.dart';
+
+/// A scrollable grid of years to allow picking a year.
+///
+/// The year picker widget is rarely used directly. Instead, consider using
+/// [CalendarDatePicker2], or [showDatePicker2] which create full date pickers.
+///
+/// See also:
+///
+///  * [CalendarDatePicker2], which provides a Material Design date picker
+///    interface.
+///
+///  * [showDatePicker2], which shows a dialog containing a Material Design
+///    date picker.
+///
+class C2YearPicker extends StatefulWidget {
+  /// Creates a year picker.
+  const C2YearPicker({
+    required this.config,
+    required this.selectedDates,
+    required this.onChanged,
+    required this.initialMonth,
+    this.dragStartBehavior = DragStartBehavior.start,
+    Key? key,
+  }) : super(key: key);
+
+  /// The calendar configurations
+  final CalendarDatePicker2Config config;
+
+  /// The currently selected dates.
+  ///
+  /// Selected dates are highlighted in the picker.
+  final List<DateTime?> selectedDates;
+
+  /// Called when the user picks a year.
+  final ValueChanged<DateTime> onChanged;
+
+  /// The initial month to display.
+  final DateTime initialMonth;
+
+  /// {@macro flutter.widgets.scrollable.dragStartBehavior}
+  final DragStartBehavior dragStartBehavior;
+
+  @override
+  State<C2YearPicker> createState() => _C2YearPickerState();
+}
+
+class _C2YearPickerState extends State<C2YearPicker> {
+  late ScrollController _scrollController;
+
+  // The approximate number of years necessary to fill the available space.
+  static const int minYears = 18;
+
+  @override
+  void initState() {
+    super.initState();
+    final scrollOffset =
+        widget.selectedDates.isNotEmpty && widget.selectedDates[0] != null
+            ? _scrollOffsetForYear(widget.selectedDates[0]!)
+            : _scrollOffsetForYear(DateUtils.dateOnly(DateTime.now()));
+    _scrollController = ScrollController(initialScrollOffset: scrollOffset);
+  }
+
+  @override
+  void didUpdateWidget(C2YearPicker oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.selectedDates != oldWidget.selectedDates) {
+      final scrollOffset =
+          widget.selectedDates.isNotEmpty && widget.selectedDates[0] != null
+              ? _scrollOffsetForYear(widget.selectedDates[0]!)
+              : _scrollOffsetForYear(DateUtils.dateOnly(DateTime.now()));
+      _scrollController.jumpTo(scrollOffset);
+    }
+  }
+
+  double _scrollOffsetForYear(DateTime date) {
+    final int initialYearIndex = date.year - widget.config.firstDate.year;
+    final int initialYearRow =
+        initialYearIndex ~/ C2Constants.yearPickerColumnCount;
+    // Move the offset down by 2 rows to approximately center it.
+    final int centeredYearRow = initialYearRow - 2;
+    return _itemCount < minYears
+        ? 0
+        : centeredYearRow * C2Constants.yearPickerRowHeight;
+  }
+
+  Widget _buildYearItem(BuildContext context, int index) {
+    final ColorScheme colorScheme = Theme.of(context).colorScheme;
+    final TextTheme textTheme = Theme.of(context).textTheme;
+
+    // Backfill the _YearPicker with disabled years if necessary.
+    final int offset = _itemCount < minYears ? (minYears - _itemCount) ~/ 2 : 0;
+    final int year = widget.config.firstDate.year + index - offset;
+    final bool isSelected = widget.selectedDates.any((d) => d?.year == year);
+    final bool isCurrentYear = year == widget.config.currentDate.year;
+    final bool isDisabled = year < widget.config.firstDate.year ||
+        year > widget.config.lastDate.year;
+    const double decorationHeight = 36.0;
+    const double decorationWidth = 72.0;
+
+    final Color textColor;
+    if (isSelected) {
+      textColor = colorScheme.onPrimary;
+    } else if (isDisabled) {
+      textColor = colorScheme.onSurface.withOpacity(0.38);
+    } else if (isCurrentYear) {
+      textColor =
+          widget.config.selectedDayHighlightColor ?? colorScheme.primary;
+    } else {
+      textColor = colorScheme.onSurface.withOpacity(0.87);
+    }
+    TextStyle? itemStyle = widget.config.yearTextStyle ??
+        textTheme.bodyLarge?.apply(color: textColor);
+    if (isSelected) {
+      itemStyle = widget.config.selectedYearTextStyle ?? itemStyle;
+    }
+
+    BoxDecoration? decoration;
+    if (isSelected) {
+      decoration = BoxDecoration(
+        color: widget.config.selectedDayHighlightColor ?? colorScheme.primary,
+        borderRadius: widget.config.yearBorderRadius ??
+            BorderRadius.circular(decorationHeight / 2),
+      );
+    } else if (isCurrentYear && !isDisabled) {
+      decoration = BoxDecoration(
+        border: Border.all(
+          color: widget.config.selectedDayHighlightColor ?? colorScheme.primary,
+        ),
+        borderRadius: widget.config.yearBorderRadius ??
+            BorderRadius.circular(decorationHeight / 2),
+      );
+    }
+
+    Widget yearItem = widget.config.yearBuilder?.call(
+          year: year,
+          textStyle: itemStyle,
+          decoration: decoration,
+          isSelected: isSelected,
+          isDisabled: isDisabled,
+          isCurrentYear: isCurrentYear,
+        ) ??
+        Center(
+          child: Container(
+            decoration: decoration,
+            height: decorationHeight,
+            width: decorationWidth,
+            child: Center(
+              child: Semantics(
+                selected: isSelected,
+                button: true,
+                child: Text(
+                  year.toString(),
+                  style: itemStyle,
+                ),
+              ),
+            ),
+          ),
+        );
+
+    if (isDisabled) {
+      yearItem = ExcludeSemantics(
+        child: yearItem,
+      );
+    } else {
+      yearItem = InkWell(
+        key: ValueKey<int>(year),
+        onTap: () => widget.onChanged(
+          DateTime(
+            year,
+            widget.initialMonth.month,
+          ),
+        ),
+        child: yearItem,
+      );
+    }
+
+    return yearItem;
+  }
+
+  int get _itemCount {
+    return widget.config.lastDate.year - widget.config.firstDate.year + 1;
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    assert(debugCheckHasMaterial(context));
+    return Column(
+      children: <Widget>[
+        const Divider(),
+        Expanded(
+          child: GridView.builder(
+            controller: _scrollController,
+            dragStartBehavior: widget.dragStartBehavior,
+            gridDelegate: _yearPickerGridDelegate,
+            itemBuilder: _buildYearItem,
+            itemCount: math.max(_itemCount, minYears),
+            padding: const EdgeInsets.symmetric(
+              horizontal: C2Constants.yearPickerPadding,
+            ),
+          ),
+        ),
+        const Divider(),
+      ],
+    );
+  }
+}
+
+class _YearPickerGridDelegate extends SliverGridDelegate {
+  const _YearPickerGridDelegate();
+
+  @override
+  SliverGridLayout getLayout(SliverConstraints constraints) {
+    final double tileWidth = (constraints.crossAxisExtent -
+            (C2Constants.yearPickerColumnCount - 1) *
+                C2Constants.yearPickerRowSpacing) /
+        C2Constants.yearPickerColumnCount;
+    return SliverGridRegularTileLayout(
+      childCrossAxisExtent: tileWidth,
+      childMainAxisExtent: C2Constants.yearPickerRowHeight,
+      crossAxisCount: C2Constants.yearPickerColumnCount,
+      crossAxisStride: tileWidth + C2Constants.yearPickerRowSpacing,
+      mainAxisStride: C2Constants.yearPickerRowHeight,
+      reverseCrossAxis: axisDirectionIsReversed(constraints.crossAxisDirection),
+    );
+  }
+
+  @override
+  bool shouldRelayout(_YearPickerGridDelegate oldDelegate) => false;
+}
+
+const _YearPickerGridDelegate _yearPickerGridDelegate =
+    _YearPickerGridDelegate();
